requirements.txt
tensorflow
pandas
numpy
scikit-learn
google-cloud-bigquery